+++ variants/models.py
@@ -169,6 +169,8 @@ class SmallVariant(models.Model):
             GinIndex(fields=["case_id", "refseq_effect"]),
             GinIndex(fields=["case_id", "ensembl_effect"]),
             models.Index(fields=["case_id", "in_clinvar"]),
+            # Fast white-list queries of gene.
+            models.Index(fields=["case_id", "ensembl_gene_id", "refseq_gene_id"]),
         ]